{"schema":"decodex/commit/1","summary":"Split operator autonomy status readback","authority":"XY-1115"}
